+++ sdk/securityinsight/arm-securityinsight/src/operations/index.ts
@@ -40,3 +40,4 @@ export * from "./watchlistItems";
 export * from "./dataConnectors";
 export * from "./dataConnectorsCheckRequirementsOperations";
 export * from "./operations";
+export * from "./summaries";